Create Bubble_sort.py
Bubble Sort(also known as sinking sort), is a simple sorting algorithm which sorts n number of elements in the list by comparing the each pair of items and swaps them if they are in the wrong order.

Following are the steps involved in bubble sort(for sorting a given array in ascending order):

Starting with the first element(index = 0), compare the current element with the next element of the array.
If the current element is greater than the next element of the array, swap them.
If the current element is less than the next element, move to the next element. Repeat Step 1.